Reorganize directory structure b/c of future changes
- introduce app.events/
- introduce app.vendors/
