+++ Dockerfile
@@ -4,9 +4,8 @@ RUN apk --no-cache upgrade && \
     apk --no-cache add github-cli py3-pip && \
     rm -rf /var/cache/apk/*
 RUN /usr/bin/python3 -m venv /opt/venv && \
-    mkdir /opt/app && \
     chown -R 1001 /opt/*
-COPY --chown=1001:root ./app/* /opt/app/
+COPY --chown=1001:root ./app/ /opt/app
 
 ENV PATH=/opt/venv/bin:${PATH} \
     VIRTUAL_ENV=/opt/venv
