+++ app/vendors/github.py
@@ -1,5 +1,5 @@
 """
-Load payload from GitHub, GitLab, Azure DevOps, etc
+GitHub specific functionality!
 """
 
 # pylint: disable=missing-function-docstring
@@ -9,31 +9,12 @@ Load payload from GitHub, GitLab, Azure DevOps, etc
 # Licensed under the AGPL-3.0: https://www.gnu.org/licenses/agpl-3.0.html
 #
 
-import json
 import os
 
 import github
 import github.Auth
 import github.IssueComment
-
-
-def load():
-    """
-    Sanity check the calling environment and load event payload.
-    """
-    if os.environ.get("GITHUB_ACTIONS") == "true":
-        return GitHubEvent(os.environ["GITHUB_EVENT_PATH"])
-
-    raise RuntimeError("Unsupported environment")
-
-
-class TriggerEvent:
-    argv = []
-    private = True
-
-    def __init__(self, file_path):
-        with open(file_path, "r", encoding="utf-8") as event_file:
-            self.payload = json.loads(event_file.read())
+from app.vendors import TriggerEvent
 
 
 class GitHubEvent(TriggerEvent):
